pyproject.toml
[project]
name = "docu-gen"
version = "0.1.0"
description = "AI-powered documentation generation system using Google Gemini API and AutoGen"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "DOCU-GEN Team"}
]
license = {text = "MIT"}

dependencies = [
    "google-generativeai>=0.3.0",
    "ratelimit>=2.2.1",
    "diskcache>=5.6.3",
    "textstat>=0.7.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
full = [
    "pyautogen>=0.2.0",
    "langchain>=0.1.0",
    "jinja2>=3.1.2",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = ["src/**/*.py"]
exclude = ["**/__pycache__", "**/*.pyc"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

